scheduler/requirements.txt
boto3==1.26.161
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-api-python-client==2.86.0
google-auth==2.24.0
pytz==2024.1
